Emmagatzematge de dades
Emmagatzematge de dades de steps i health connect

+++ FrontendTFG/src/screens/SetDailyObjectiveScreen.js
@@ -22,7 +22,7 @@ const SetDailyObjectiveScreen = ({ navigation }) => {
 
     try {
       setIsSubmitting(true);
-      await addDailyObjective(token, parseFloat(dailyStepsObjective));
+      await addDailyObjective(token, parseInt(dailyStepsObjective));
       
       setIsNewUser(false);
